WIP - customizable settings dropdown
